refactor(search applications): rename engine error file
