+++ x-pack/plugins/enterprise_search/public/applications/applications/components/search_application/search_application_content.tsx
@@ -25,10 +25,10 @@ import {
 import { EnterpriseSearchApplicationsPageTemplate } from '../layout/page_template';
 
 import { AddIndicesFlyout } from './add_indices_flyout';
-import { EngineError } from './engine_error';
 import { EngineIndices } from './engine_indices';
 import { EngineIndicesLogic } from './engine_indices_logic';
 import { EngineSchema } from './engine_schema';
+import { EngineError } from './search_application_error';
 import { SearchApplicationViewLogic } from './search_application_view_logic';
 import './search_application_layout.scss';
