+++ x-pack/plugins/enterprise_search/public/applications/applications/components/search_applications/search_application_indices_flyout.tsx
@@ -35,7 +35,7 @@ import { healthColorsMap } from '../../../shared/constants/health_colors';
 import { generateEncodedPath } from '../../../shared/encode_path_params';
 import { EuiLinkTo } from '../../../shared/react_router_helpers';
 
-import { EngineError } from '../search_application/engine_error';
+import { EngineError } from '../search_application/search_application_error';
 
 import { SearchApplicationIndicesFlyoutLogic } from './search_application_indices_flyout_logic';
